Add db commits to fixtures to make ctl tests happy
